backend/requirements.txt
Flask
flask-cors
python-dotenv
supabase
google-generativeai
Pillowpip install -r requirements.txt